requirements.txt
allure-pytest==2.8.32
pytest-html==3.1.1
requests==2.25.1
selenium==3.141.0
